Rename embed.js after class name
